Closed #45; looks like just converting the values to a float works just fine.

+++ dataview/views.py
@@ -120,7 +120,8 @@ def dump_graph_data(request, sub_id):
     usages = Usage.objects.filter(sub=sub_id)[:20]   # pk or usage_sub_id?
 
     for use in usages:
-        usage_graph_data[str(use.timestamp)] = int(use.dosage)
+        usage_graph_data[str(use.timestamp)] = float(use.dosage)
+        # TODO: find a workaround for this only working with integers
         # later on we can look at using use.notes as hover-over text for each
         # graph bar, or something of the like
